Split ChangeSlabSize code into a directory
This part of the algorithm has a lot connected to it, and too much was
ending up in large files.

Only content change is moving the tags out of a detail namespace.
