+++ src/Evolution/Deadlock/PrintDgElementArray.hpp
@@ -15,7 +15,7 @@
 #include "Parallel/GlobalCache.hpp"
 #include "Parallel/OutputInbox.hpp"
 #include "Parallel/Printf.hpp"
-#include "Time/Actions/ChangeSlabSize.hpp"
+#include "Time/ChangeSlabSize/Tags.hpp"
 #include "Time/Tags/Time.hpp"
 #include "Time/Tags/TimeStep.hpp"
 #include "Time/Tags/TimeStepId.hpp"
@@ -36,7 +36,7 @@ namespace deadlock {
  * \details This will print the contents of the following inbox or DataBox tags:
  *
  * - `evolution::dg::Tags::BoundaryCorrectionAndGhostCellsInbox<3>`
- * - `ChangeSlabSize_detail::NewSlabSizeInbox`
+ * - `Tags::ChangeSlabSize::NewSlabSizeInbox`
  * - `evolution::dg::Tags::MortarNextTemporalId<3>`
  * - `evolution::dg::Tags::MortarDataHistory` (for LTS only)
  * - `evolution::dg::Tags::MortarData<3>` (for GTS only)
@@ -88,7 +88,7 @@ struct PrintElementInfo {
           evolution::dg::Tags::BoundaryCorrectionAndGhostCellsInbox<3>>(inboxes,
                                                                         2_st);
       const std::string slab_size_inbox =
-          Parallel::output_inbox<ChangeSlabSize_detail::NewSlabSizeInbox>(
+          Parallel::output_inbox<::Tags::ChangeSlabSize::NewSlabSizeInbox>(
               inboxes, 2_st);
       ss << mortar_inbox;
       ss << slab_size_inbox;
