+++ src/Evolution/Executables/GeneralizedHarmonic/EvolveGhAndCharacteristic.hpp
@@ -50,6 +50,7 @@
 #include "ParallelAlgorithms/Interpolation/Protocols/InterpolationTargetTag.hpp"
 #include "ParallelAlgorithms/Interpolation/Targets/KerrHorizon.hpp"
 #include "ParallelAlgorithms/Interpolation/Targets/Sphere.hpp"
+#include "Time/ChangeSlabSize/Action.hpp"
 #include "Time/Tags/TimeStepId.hpp"
 #include "Utilities/ErrorHandling/Error.hpp"
 #include "Utilities/ProtocolHelpers.hpp"
