+++ src/Time/ChangeSlabSize/CMakeLists.txt
@@ -1,13 +1,18 @@
 # Distributed under the MIT License.
 # See LICENSE.txt for details.
 
+spectre_target_sources(
+  ${LIBRARY}
+  PRIVATE
+  Event.cpp
+  )
+
 spectre_target_headers(
   ${LIBRARY}
   INCLUDE_DIRECTORY ${CMAKE_SOURCE_DIR}/src
   HEADERS
-  AdvanceTime.hpp
-  ChangeStepSize.hpp
-  RecordTimeStepperData.hpp
-  SelfStartActions.hpp
-  UpdateU.hpp
+  Action.hpp
+  ChangeSlabSize.hpp
+  Event.hpp
+  Tags.hpp
   )
