+++ src/Time/ChangeSlabSize/Event.cpp
@@ -1,7 +1,7 @@
 // Distributed under the MIT License.
 // See LICENSE.txt for details.
 
-#include "Time/Actions/ChangeSlabSize.hpp"
+#include "Time/ChangeSlabSize/Event.hpp"
 
 namespace Events {
 PUP::able::PUP_ID ChangeSlabSize::my_PUP_ID = 0;  // NOLINT
