+++ src/Time/ChangeSlabSize/Event.hpp
@@ -4,104 +4,38 @@
 #pragma once
 
 #include <algorithm>
-#include <boost/functional/hash.hpp>
 #include <cstddef>
 #include <cstdint>
-#include <iomanip>
 #include <limits>
-#include <map>
 #include <memory>
-#include <optional>
 #include <pup.h>
 #include <pup_stl.h>
-#include <sstream>
-#include <string>
-#include <tuple>
-#include <unordered_set>
 #include <utility>
 #include <vector>
 
 #include "DataStructures/DataBox/DataBox.hpp"
 #include "Options/String.hpp"
-#include "Parallel/AlgorithmExecution.hpp"
 #include "Parallel/GlobalCache.hpp"
-#include "Parallel/InboxInserters.hpp"
 #include "Parallel/Invoke.hpp"
 #include "Parallel/Local.hpp"
 #include "Parallel/Reduction.hpp"
 #include "ParallelAlgorithms/EventsAndTriggers/Event.hpp"
-#include "Time/ChangeSlabSize.hpp"
+#include "Time/ChangeSlabSize/Tags.hpp"
 #include "Time/StepChoosers/StepChooser.hpp"
-#include "Time/Tags/HistoryEvolvedVariables.hpp"
 #include "Time/TimeStepId.hpp"
-#include "Time/TimeSteppers/TimeStepper.hpp"
-#include "Utilities/Algorithm.hpp"
-#include "Utilities/ErrorHandling/Assert.hpp"
 #include "Utilities/Functional.hpp"
 #include "Utilities/Serialization/CharmPupable.hpp"
 #include "Utilities/TMPL.hpp"
-#include "Utilities/TaggedTuple.hpp"
 
 /// \cond
 namespace Tags {
 struct DataBox;
 struct TimeStepId;
-template <typename StepperInterface>
-struct TimeStepper;
 }  // namespace Tags
 /// \endcond
 
+namespace Events {
 namespace ChangeSlabSize_detail {
-struct NewSlabSizeInbox
-    : public Parallel::InboxInserters::MemberInsert<NewSlabSizeInbox> {
-  using temporal_id = int64_t;
-  using type = std::map<temporal_id, std::unordered_multiset<double>>;
-
-  static std::string output_inbox(const type& inbox,
-                                  const size_t padding_size) {
-    std::stringstream ss{};
-    const std::string pad(padding_size, ' ');
-
-    ss << std::scientific << std::setprecision(16);
-    ss << pad << "NewSlabSizeInbox:\n";
-    for (const auto& [slab_number, slab_sizes] : inbox) {
-      ss << pad << " Slab number: " << slab_number
-         << ", slab sizes: " << slab_sizes << "\n";
-    }
-
-    return ss.str();
-  }
-};
-
-// This inbox doesn't receive any data, it just counts messages (using
-// the size of the multiset).  Whenever a message is sent to the
-// NewSlabSizeInbox, another message is sent here synchronously, so
-// the count here is the number of messages we expect in the
-// NewSlabSizeInbox.
-struct NumberOfExpectedMessagesInbox
-    : public Parallel::InboxInserters::MemberInsert<
-          NumberOfExpectedMessagesInbox> {
-  using temporal_id = int64_t;
-  using NoData = std::tuple<>;
-  using type = std::map<temporal_id,
-                        std::unordered_multiset<NoData, boost::hash<NoData>>>;
-
-  static std::string output_inbox(const type& inbox,
-                                  const size_t padding_size) {
-    std::stringstream ss{};
-    const std::string pad(padding_size, ' ');
-
-    ss << std::scientific << std::setprecision(16);
-    ss << pad << "SlabSizeNumberOfExpectedMessagesInbox:\n";
-    for (const auto& [number_of_messages, no_data] : inbox) {
-      (void)no_data;
-      ss << pad << " Number of messages: " << number_of_messages << "\n";
-    }
-
-    return ss.str();
-  }
-};
-
 struct StoreNewSlabSize {
   template <typename ParallelComponent, typename DbTags, typename Metavariables,
             typename ArrayIndex>
@@ -109,7 +43,7 @@ struct StoreNewSlabSize {
                     Parallel::GlobalCache<Metavariables>& cache,
                     const ArrayIndex& array_index, const int64_t slab_number,
                     const double slab_size) {
-    Parallel::receive_data<ChangeSlabSize_detail::NewSlabSizeInbox>(
+    Parallel::receive_data<::Tags::ChangeSlabSize::NewSlabSizeInbox>(
         *Parallel::local(Parallel::get_parallel_component<ParallelComponent>(
             cache)[array_index]),
         slab_number, slab_size);
@@ -117,111 +51,6 @@ struct StoreNewSlabSize {
 };
 }  // namespace ChangeSlabSize_detail
 
-namespace Actions {
-/// \ingroup ActionsGroup
-/// \ingroup TimeGroup
-/// Adjust the slab size based on previous executions of
-/// Events::ChangeSlabSize
-///
-/// Uses:
-/// - DataBox:
-///   - Tags::HistoryEvolvedVariables
-///   - Tags::TimeStep
-///   - Tags::TimeStepId
-///   - Tags::TimeStepper<TimeStepper>
-///
-/// DataBox changes:
-/// - Adds: nothing
-/// - Removes: nothing
-/// - Modifies:
-///   - Tags::Next<Tags::TimeStepId>
-///   - Tags::TimeStep
-///   - Tags::TimeStepId
-struct ChangeSlabSize {
-  using inbox_tags =
-      tmpl::list<ChangeSlabSize_detail::NumberOfExpectedMessagesInbox,
-                 ChangeSlabSize_detail::NewSlabSizeInbox>;
-
-  template <typename DbTags, typename... InboxTags, typename Metavariables,
-            typename ArrayIndex, typename ActionList,
-            typename ParallelComponent>
-  static Parallel::iterable_action_return_t apply(
-      db::DataBox<DbTags>& box, tuples::TaggedTuple<InboxTags...>& inboxes,
-      const Parallel::GlobalCache<Metavariables>& /*cache*/,
-      const ArrayIndex& /*array_index*/, const ActionList /*meta*/,
-      const ParallelComponent* const /*meta*/) {
-    const auto& time_step_id = db::get<::Tags::TimeStepId>(box);
-    if (not time_step_id.is_at_slab_boundary()) {
-      return {Parallel::AlgorithmExecution::Continue, std::nullopt};
-    }
-
-    auto& message_count_inbox =
-        tuples::get<ChangeSlabSize_detail::NumberOfExpectedMessagesInbox>(
-            inboxes);
-    if (message_count_inbox.empty() or
-        message_count_inbox.begin()->first != time_step_id.slab_number()) {
-      return {Parallel::AlgorithmExecution::Continue, std::nullopt};
-    }
-
-    auto& new_slab_size_inbox =
-        tuples::get<ChangeSlabSize_detail::NewSlabSizeInbox>(inboxes);
-
-    const auto slab_number = time_step_id.slab_number();
-    const auto number_of_changes = [&slab_number](const auto& inbox) -> size_t {
-      if (inbox.empty()) {
-        return 0;
-      }
-      if (inbox.begin()->first == slab_number) {
-        return inbox.begin()->second.size();
-      }
-      ASSERT(inbox.begin()->first >= slab_number,
-             "Received data for a change at slab " << inbox.begin()->first
-             << " but it is already slab " << slab_number);
-      return 0;
-    };
-
-    const size_t expected_messages = number_of_changes(message_count_inbox);
-    const size_t received_messages = number_of_changes(new_slab_size_inbox);
-    ASSERT(expected_messages >= received_messages,
-           "Received " << received_messages << " size change messages at slab "
-                       << slab_number << ", but only expected "
-                       << expected_messages);
-    if (expected_messages != received_messages) {
-      return {Parallel::AlgorithmExecution::Retry, std::nullopt};
-    }
-
-    message_count_inbox.erase(message_count_inbox.begin());
-
-    const double new_slab_size =
-        *alg::min_element(new_slab_size_inbox.begin()->second);
-    new_slab_size_inbox.erase(new_slab_size_inbox.begin());
-
-    const TimeStepper& time_stepper =
-        db::get<::Tags::TimeStepper<TimeStepper>>(box);
-
-    // Sometimes time steppers need to run with a fixed step size.
-    // This is generally at the start of an evolution when the history
-    // is in an unusual state.
-    if (not time_stepper.can_change_step_size(
-            time_step_id, db::get<::Tags::HistoryEvolvedVariables<>>(box))) {
-      return {Parallel::AlgorithmExecution::Continue, std::nullopt};
-    }
-
-    const auto current_slab = time_step_id.step_time().slab();
-
-    const double new_slab_end =
-        time_step_id.time_runs_forward()
-            ? current_slab.start().value() + new_slab_size
-            : current_slab.end().value() - new_slab_size;
-
-    change_slab_size(make_not_null(&box), new_slab_end);
-
-    return {Parallel::AlgorithmExecution::Continue, std::nullopt};
-  }
-};
-}  // namespace Actions
-
-namespace Events {
 /// \ingroup TimeGroup
 /// %Trigger a slab size change.
 ///
@@ -314,16 +143,16 @@ class ChangeSlabSize : public Event {
     // This message is sent synchronously, so it is guaranteed to
     // arrive before the ChangeSlabSize action is called.
     Parallel::receive_data<
-        ChangeSlabSize_detail::NumberOfExpectedMessagesInbox>(
+        ::Tags::ChangeSlabSize::NumberOfExpectedMessagesInbox>(
         *Parallel::local(self_proxy), slab_to_change,
-        ChangeSlabSize_detail::NumberOfExpectedMessagesInbox::NoData{});
+        ::Tags::ChangeSlabSize::NumberOfExpectedMessagesInbox::NoData{});
     if (synchronization_required) {
       Parallel::contribute_to_reduction<
           ChangeSlabSize_detail::StoreNewSlabSize>(
           ReductionData(slab_to_change, desired_slab_size), self_proxy,
           component_proxy);
     } else {
-      Parallel::receive_data<ChangeSlabSize_detail::NewSlabSizeInbox>(
+      Parallel::receive_data<::Tags::ChangeSlabSize::NewSlabSizeInbox>(
           *Parallel::local(self_proxy), slab_to_change, desired_slab_size);
     }
   }
