+++ tests/Unit/Time/Actions/CMakeLists.txt
@@ -4,7 +4,6 @@
 set(LIBRARY_SOURCES
   ${LIBRARY_SOURCES}
   Actions/Test_AdvanceTime.cpp
-  Actions/Test_ChangeSlabSize.cpp
   Actions/Test_ChangeStepSize.cpp
   Actions/Test_RecordTimeStepperData.cpp
   Actions/Test_SelfStartActions.cpp
