+++ tests/Unit/Time/ChangeSlabSize/CMakeLists.txt
@@ -1,0 +1,8 @@
+# Distributed under the MIT License.
+# See LICENSE.txt for details.
+
+set(LIBRARY_SOURCES
+  ${LIBRARY_SOURCES}
+  ChangeSlabSize/Test_Action.cpp
+  ChangeSlabSize/Test_ChangeSlabSize.cpp
+  PARENT_SCOPE)
